Separate MapRenderer utils.
